requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
pyChomikBox==0.2.0
requests==2.23.0
requests-toolbelt==0.9.1
urllib3==1.25.8
xmltodict==0.12.0
